// algamoney-ui/src/app/services/models/page-response-pessoa-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PessoaResponseDto } from '../models/pessoa-response-dto';
export interface PageResponsePessoaResponseDto {
  content?: Array<PessoaResponseDto>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}